requirements.txt
Python >= 2.6
PyCogent >= 1.5
Numpy >= 1.3.0
Matplotlib >= 0.98.3
